+++ src/scene_object_manager.cpp
@@ -36,8 +36,8 @@ void SceneObjectManager::BuildWorld() {
           (*objectPtr)->Update();
 
           // Insert the object into the space partition tree used for object collision detection
-          std::vector<uint16_t> lowerBound = (*objectPtr)->GetLowerBound();
-          std::vector<uint16_t> upperBound = (*objectPtr)->GetUpperBound();
+          std::vector<int> lowerBound = (*objectPtr)->GetLowerBound();
+          std::vector<int> upperBound = (*objectPtr)->GetUpperBound();
           spacePartitionObjectsTree->insertParticle(*objectPtr, lowerBound, upperBound);
 
           // Save pointers to proper arrays for static objects and mobile objects
